+++ app/models/runn_leave.rb
@@ -1,3 +1,3 @@
-class RunnLeaveMirror < ApplicationRecord
+class RunnLeave < ApplicationRecord
   validates :runn_person_id, :start_date, :end_date, :refreshed_at, presence: true
 end
