+++ app/services/mcp/server.rb
@@ -21,7 +21,6 @@ module Mcp
       Mcp::ListOpenAdminTasksTool,
       Mcp::ListProjectsAtRiskTool,
       Mcp::GetStudioHealthTool,
-      Mcp::GetResourcingWorldTool,
       Mcp::GetResourcingProjectionsTool,
     ].freeze
