+++ app/services/resourcing/runn_leave_refresh.rb
@@ -1,7 +1,7 @@
 module Resourcing
-  # Nightly mirror of Runn per-person leave into runn_leave_mirrors so the
-  # get_resourcing_world read never makes N live per-person calls on the request path.
-  class LeaveMirrorRefresh
+  # Nightly mirror of Runn per-person leave into runn_leaves so the
+  # get_resourcing_projections read never makes N live per-person calls on the request path.
+  class RunnLeaveRefresh
     def initialize(runn: Stacks::Runn.new(max_retries: 5))
       @runn = runn
     end
@@ -13,10 +13,10 @@ module Resourcing
         pid = person["id"]
         begin
           periods = @runn.get_leave_for_person(pid)
-          RunnLeaveMirror.transaction do
-            RunnLeaveMirror.where(runn_person_id: pid).delete_all
+          RunnLeave.transaction do
+            RunnLeave.where(runn_person_id: pid).delete_all
             periods.each do |lp|
-              RunnLeaveMirror.create!(
+              RunnLeave.create!(
                 runn_person_id: pid,
                 start_date: Date.parse(lp["startDate"].to_s),
                 end_date: Date.parse(lp["endDate"].to_s),
