+++ lib/tasks/resourcing.rake
@@ -1,7 +1,7 @@
 namespace :resourcing do
-  desc "Mirror Runn per-person leave into runn_leave_mirrors (nightly)"
-  task refresh_leave_mirror: :environment do
-    count = Resourcing::LeaveMirrorRefresh.new.run!
-    puts "[resourcing] refreshed leave mirror for #{count} people"
+  desc "Mirror Runn per-person leave into runn_leaves (nightly)"
+  task refresh_runn_leave: :environment do
+    count = Resourcing::RunnLeaveRefresh.new.run!
+    puts "[resourcing] refreshed runn leave for #{count} people"
   end
 end
