+++ test/services/resourcing/runn_leave_refresh_test.rb
@@ -1,7 +1,7 @@
 require "test_helper"
 
-class Resourcing::LeaveMirrorRefreshTest < ActiveSupport::TestCase
-  test "run! mirrors each non-archived person's leave into runn_leave_mirrors" do
+class Resourcing::RunnLeaveRefreshTest < ActiveSupport::TestCase
+  test "run! mirrors each non-archived person's leave into runn_leaves" do
     runn = mock("runn")
     runn.stubs(:get_people).returns([
       { "id" => 10, "isArchived" => false },
@@ -12,34 +12,34 @@ class Resourcing::LeaveMirrorRefreshTest < ActiveSupport::TestCase
     ])
     runn.expects(:get_leave_for_person).with(11).never
 
-    svc = Resourcing::LeaveMirrorRefresh.new(runn: runn)
+    svc = Resourcing::RunnLeaveRefresh.new(runn: runn)
     count = svc.run!
     assert_equal 1, count
-    row = RunnLeaveMirror.find_by(runn_person_id: 10)
+    row = RunnLeave.find_by(runn_person_id: 10)
     assert_equal Date.new(2030, 5, 10), row.start_date
     assert row.refreshed_at.present?
   end
 
   test "run! replaces a person's prior leave rows" do
-    RunnLeaveMirror.create!(runn_person_id: 10, start_date: Date.new(2029, 1, 1), end_date: Date.new(2029, 1, 2), refreshed_at: Time.current)
+    RunnLeave.create!(runn_person_id: 10, start_date: Date.new(2029, 1, 1), end_date: Date.new(2029, 1, 2), refreshed_at: Time.current)
     runn = mock("runn")
     runn.stubs(:get_people).returns([{ "id" => 10, "isArchived" => false }])
     runn.stubs(:get_leave_for_person).returns([{ "startDate" => "2030-05-10", "endDate" => "2030-05-15" }])
-    svc = Resourcing::LeaveMirrorRefresh.new(runn: runn)
+    svc = Resourcing::RunnLeaveRefresh.new(runn: runn)
     svc.run!
-    assert_equal [Date.new(2030, 5, 10)], RunnLeaveMirror.where(runn_person_id: 10).pluck(:start_date)
+    assert_equal [Date.new(2030, 5, 10)], RunnLeave.where(runn_person_id: 10).pluck(:start_date)
   end
 
   test "run! isolates a per-person failure and still mirrors the rest" do
-    RunnLeaveMirror.create!(runn_person_id: 11, start_date: Date.new(2029, 1, 1), end_date: Date.new(2029, 1, 2),
+    RunnLeave.create!(runn_person_id: 11, start_date: Date.new(2029, 1, 1), end_date: Date.new(2029, 1, 2),
       refreshed_at: 1.week.ago)
     runn = mock("runn")
     runn.stubs(:get_people).returns([{ "id" => 10, "isArchived" => false }, { "id" => 11, "isArchived" => false }])
     runn.stubs(:get_leave_for_person).with(10).returns([{ "startDate" => "2030-05-10", "endDate" => "2030-05-15" }])
     runn.stubs(:get_leave_for_person).with(11).raises(RuntimeError, "runn 500")
-    svc = Resourcing::LeaveMirrorRefresh.new(runn: runn)
+    svc = Resourcing::RunnLeaveRefresh.new(runn: runn)
     assert_nothing_raised { svc.run! }
-    assert RunnLeaveMirror.find_by(runn_person_id: 10).present?
-    assert_equal Date.new(2029, 1, 1), RunnLeaveMirror.find_by(runn_person_id: 11).start_date
+    assert RunnLeave.find_by(runn_person_id: 10).present?
+    assert_equal Date.new(2029, 1, 1), RunnLeave.find_by(runn_person_id: 11).start_date
   end
 end
